+++ src/main/java/net/fyoncle/elysiumcore/utility/other/Strings.java
@@ -2,6 +2,4 @@ package net.fyoncle.elysiumcore.utility.other;
 
 public class Strings {
     public static String LATEST_ED_VERSION = "0.0.0";
-    public static String HEALTH_BAR_TOGGLED_STATE = "Health Bar: ON";
-    public static String HEALTH_BAR_UNTOGGLED_STATE = "Health Bar: OFF";
 }